Tried backend with direct links, causes some issues due to lack of parsing. We need a good news API that gives us the raw article text?
